requirements.txt
google-adk
python-dotenv
pymupdf
pytesseract
Pillow
google-genai
